refactor(providers): move the managed-soniox split module beside its future caller
