+++ src/services/providers/managedSonioxSplit.ts
@@ -1,11 +1,11 @@
 import type {
   ManagedSonioxSession,
   SonioxSessionMatrixInput,
-} from '../../services/clients/ManagedSonioxSession';
+} from '../clients/ManagedSonioxSession';
 // Type-only, like the import above: this keeps the module free of any runtime
 // edge to the descriptor layer (and the client graph behind it), which matters
 // because the subtitle window loads MainPanel's helpers too.
-import type { ClientOptions } from '../../services/providers/ProviderDescriptor';
+import type { ClientOptions } from './ProviderDescriptor';
 
 /**
  * Pure wiring decisions for a managed (Kizuna AI) Soniox session, extracted out
@@ -16,6 +16,8 @@ import type { ClientOptions } from '../../services/providers/ProviderDescriptor'
  * (managedVoicePrep.ts) follows, and for the same reason: the alternative is
  * a hand-transcribed duplicate inside a test that drifts from the real branch
  * without either side noticing. Only the side effects stay inline in MainPanel.
+ *
+ * Moved beside its caller (KizunaAISonioxProviderConfig.acquireSessionResources, S6); stays type-only-import pure so a descriptor uses it without cycles.
  */
 
 export interface ManagedSonioxWiring {
